refactor: put attribute changer to file
